+++ changelog.md
@@ -1,7 +1,7 @@
 # 1.2.7
 
 - Added Hitbox Trail
-- Added No Wave Pulse
+- Added Quests in Pause
 
 # 1.2.6